// client/src/types.ts
export interface User {
  id: string;
  name: string;
  color: string;
  cursor?: {
    x: number;
    y: number;
    element?: "whiteboard" | "editor";
    line?: number;
    ch?: number;
  };
}

export interface Point {
  x: number;
  y: number;
}

export interface Stroke {
  id: string;
  points: Point[]; // for shapes, points[0] is start, points[1] is end
  color: string;
  width: number;
  tool: "pen" | "eraser" | "circle" | "rectangle";
  userId: string;
  userName: string;
}

export interface ActivityLog {
  id: string;
  timestamp: string;
  userName: string;
  userColor: string;
  text: string;
}

export type CodeLanguage = "javascript" | "python" | "html" | "css";